Removed Debugger.closeNodeTo
We didn't use this very often, and it was only useful for "cleaning up"
after some internal method failed to call closeNode. However, we don't
have any problems like that, and we can also ensure that closeNode is
called via a try/finally. As a result, we don't really need this.