Removed some extraneous filter options
These options were removed primarily to simplify my renovation work.
Most have legitimate use-cases, but I'd prefer their behavior to be
implemented outside of the UI classes themselves.

"Create Listener" and "Send to Filter" seem to me to be the same thing,
so I'd prefer only one of them to exist.

"Capturing" is a neat idea, but I don't really care for it. It could be
implemented as a inverse filter on a parent if I wanted to readd that
behavior.

"Refresh" and "Synchronized" are not useful. I'd prefer logs to always
be synchronized, which removes the need to have the check box or the
"Refresh" button.

I don't really know what "Default Filter" does, so I'd rather just
remove it.