Several big changes to logging
Firstly, I now use DefaultTreeModel's methods for updating the tree
model. This ensures events are generated for insertions and deletions.

Secondly, logging is now buffered such that the UI modification is done
solely on the Swing EDT. This keeps other threads responsive, as well as
ensuring we don't end up corrupting anything.

Finally, I re-implemented merging of scope groups. This isn't a bugfix,
but it greatly improves the readability of generated logs.